+++ lib/mongol.dart
@@ -3,4 +3,3 @@ library mongol;
 export 'package:mongol/mongol_text.dart';
 export 'package:mongol/mongol_rich_text.dart';
 export 'package:mongol/mongol_alert_dialog.dart';
-export 'package:mongol/mongol_font.dart';