+++ lib/mongol_font.dart
@@ -1,4 +1,0 @@
-class MongolFont {
-  static String defaultFont;
-  static const double defaultFontSize = 17.0;
-}