+++ test/mongol_text_widget_test.dart
@@ -3,8 +3,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:mongol/mongol.dart';
-import 'package:mongol/mongol_font.dart';
-import 'package:mongol/mongol_rich_text.dart';
 
 void main() {
   testWidgets('MongolText has text', (WidgetTester tester) async {
@@ -33,16 +31,4 @@ void main() {
     expect(richText, isNotNull);
     expect(richText.text.style.fontFamily, equals('Some Font'));
   });
-
-  // testWidgets('MongolText default to Mongol font', (WidgetTester tester) async {
-  //   await tester.pumpWidget(
-  //     MongolText(
-  //       'T',
-  //     ),
-  //   );
-
-  //   MongolRichText richText = tester.firstWidget(find.byType(MongolRichText));
-  //   expect(richText, isNotNull);
-  //   expect(richText.text.style.fontFamily, equals(MongolFont.qagan));
-  // });
 }